feat: 게시글 화면 기본값 삭제
