+++ core/domain/src/main/java/com/example/domain/model/post/Post.kt
@@ -51,21 +51,21 @@ data class PostFeed(
 }
 
 data class PostDetail(
-    val postId : Int,
+    val postId: Int,
     val postType: PostType,
     val viewCount: Int,
     val emotionCount: EmotionCount,
     val title: String,
     val content: String,
-    val missionContent : String? = null,
-    val providerId :String,
+    val missionContent: String? = null,
+    val providerId:String,
     val nickname: String,
     val images: List<String> = emptyList(),
     val profileImageUrl: String? = null,
-    val yourEmotionType : Emotion? = null,
+    val yourEmotionType: Emotion? = null,
     val createdAt: LocalDateTime,
     val updatedAt: LocalDateTime,
-    val isOwner : Boolean = true,
+    val isOwner: Boolean = true,
     val isVerified: Boolean = false,
 ) {
     fun getFormattedDate(): String {
